Fix change plates

+++ resources/[scripts]/garages/server-side/core.lua
@@ -522,7 +522,7 @@ AddEventHandler("garages:Spawn",function(Table)
 
 						local Network = Spawn[Plate][3]
 						local Network = NetworkGetEntityFromNetworkId(Network)
-						if DoesEntityExist(Network) and not IsPedAPlayer(Network) and GetEntityType(Network) == 2 and GetVehicleNumberPlateText(Network) == Plate then
+						if DoesEntityExist(Network) and not IsPedAPlayer(Network) and GetEntityType(Network) == 2 then
 							vCLIENT.SearchBlip(source,GetEntityCoords(Network))
 							TriggerClientEvent("Notify",source,"amarelo","Rastreador do veículo foi ativado por <b>30</b> segundos, lembrando que se o mesmo estiver em movimento a localização pode ser imprecisa.",10000)
 						else
@@ -781,10 +781,8 @@ AddEventHandler("garages:deleteVehicle",function(Network,Plate)
 		end
 
 		local Network = NetworkGetEntityFromNetworkId(Network)
-		if DoesEntityExist(Network) and not IsPedAPlayer(Network) and GetEntityType(Network) == 2 then
-			if GetVehicleNumberPlateText(Network) == Plate then
-				DeleteEntity(Network)
-			end
+		if DoesEntityExist(Network) and not IsPedAPlayer(Network) and GetEntityType(Network) == 2 and GetVehicleNumberPlateText(Network) == Plate then
+			DeleteEntity(Network)
 		end
 	end
 end)
